# pkg/api/rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Workflow management module",
        "title": "CM-Cicada REST API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "latest"
    },
    "basePath": "/cicada",
    "paths": {
...
        "/workflow/{wfId}": {
            "get": {
                "description": "Get the workflow.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Workflow]"
                ],
                "summary": "Get Workflow",
                "operationId": "get-workflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DB workflow ID.",
                        "name": "wfId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get the workflow.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.Workflow"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get the workflow.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update the workflow content.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Workflow]"
                ],
                "summary": "Update Workflow",
                "operationId": "update-workflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DB workflow ID.",
                        "name": "wfId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Workflow to modify.",
                        "name": "Workflow",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.CreateWorkflowReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully update the workflow",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.Workflow"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update the workflow",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete the workflow.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Workflow]"
                ],
                "summary": "Delete Workflow",
                "operationId": "delete-workflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DB workflow ID.",
                        "name": "wfId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully delete the workflow",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.SimpleMsg"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete the workflow",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/workflow/{wfId}/run": {
            "post": {
                "description": "Run the workflow.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Workflow]"
                ],
                "summary": "Run Workflow",
                "operationId": "run-workflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DB workflow ID.",
                        "name": "wfId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully run the workflow.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.SimpleMsg"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to run the Workflow",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workflow/{wfId}/runs": {
            "get": {
                "description": "Get the task Logs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Workflow]"
                ],
                "summary": "Get workflowRuns",
                "operationId": "get-workflow-runs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DB workflow ID.",
                        "name": "wfId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get the workflowRuns.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.WorkflowRun"
                            }
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get the workflowRuns.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workflow/{wfId}/status": {
            "get": {
                "description": "Get the WorkflowStatus.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Workflow]"
                ],
                "summary": "Get WorkflowStatus",
                "operationId": "get-WorkflowStatus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "wfId of the workflow",
                        "name": "wfId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get the WorkflowVersion.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.WorkflowStatus"
                            }
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get the WorkflowVersion.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_common.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_model.CreateDataReq": {
            "type": "object",
            "required": [
                "task_groups"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "task_groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.CreateTaskGroupReq"
                    }
                }
            }
        },
...
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_model.CreateTaskGroupReq": {
            "type": "object",
            "required": [
                "name",
                "tasks"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.CreateTaskReq"
                    }
                }
            }
        },
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_model.CreateTaskReq": {
            "type": "object",
            "required": [
                "name",
                "request_body",
                "task_component"
            ],
            "properties": {
                "dependencies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "extra": {
                    "type": "object",
                    "additionalProperties": true
                },
                "name": {
                    "type": "string"
                },
                "path_params": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "query_params": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "request_body": {
                    "type": "string"
                },
                "task_component": {
                    "type": "string"
                }
            }
        },
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_model.CreateWorkflowReq": {
            "type": "object",
            "required": [
                "data",
                "name"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.CreateDataReq"
                },
                "name": {
                    "type": "string"
                },
                "spec_version": {
                    "type": "string"
                }
            }
        },
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_model.Data": {
            "type": "object",
            "required": [
                "task_groups"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "task_groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.TaskGroup"
                    }
                }
            }
        },
...
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_model.SimpleMsg": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_model.Task": {
            "type": "object",
            "required": [
                "id",
                "name",
                "request_body",
                "task_component"
            ],
            "properties": {
                "dependencies": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "extra": {
                    "type": "object",
                    "additionalProperties": true
                },
                "id": {
                    "type": "string"
                },
                "is_deleted_task": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "path_params": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "query_params": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "request_body": {
                    "type": "string"
                },
                "task_component": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_model.TaskGroup": {
            "type": "object",
            "required": [
                "id",
                "name",
                "tasks"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.Task"
                    }
                }
            }
        },
...
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_model.Workflow": {
            "type": "object",
            "required": [
                "data",
                "id",
                "name",
                "spec_version"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/github_com_cloud-barista_cm-cicada_pkg_api_rest_model.Data"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "spec_version": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_model.WorkflowRun": {
            "type": "object",
            "properties": {
                "conf": {
                    "type": "object",
                    "additionalProperties": true
                },
                "dag_id": {
                    "type": "string"
                },
                "data_interval_end": {
                    "type": "string"
                },
                "data_interval_start": {
                    "type": "string"
                },
                "duration_date": {
                    "type": "number"
                },
                "end_date": {
                    "type": "string"
                },
                "execution_date": {
                    "type": "string"
                },
                "external_trigger": {
                    "type": "boolean"
                },
                "last_scheduling_decision": {
                    "type": "string"
                },
                "logical_date": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "run_type": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "workflow_id": {
                    "type": "string"
                },
                "workflow_run_id": {
                    "type": "string"
                }
            }
        },
        "github_com_cloud-barista_cm-cicada_pkg_api_rest_model.WorkflowStatus": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
    }
}